+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - csharp-mvc: close editor
 - java: close editor
 - php: close editor
+- php-laravel: close editor
 - he-IL skin language
 - ConvertService.ashx address replaced with converter
 - coauthoring/CommandService.ashx address replaced with command